backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sentence-transformers==4.1.0
numpy>=1.24.3
pydantic==2.11.5
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
pydantic-settings==2.2.1
sqlalchemy==2.0.28
alembic==1.13.1